// packages/runtime/src/types.ts
export type BiuAuthPageMode = "LOGIN" | "REGISTER";

export interface BiuAuthPageProps {
  mode: BiuAuthPageMode;
  onModeChange: (mode: BiuAuthPageMode) => void;
}